reflection-agent/pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["Henrik Seidel <heseber+github@mailbox.org>"]
readme = "README.md"
packages = [{include = "reflection_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
ruff = "^0.8.5"
isort = "^5.13.2"
langchain = "^0.3.13"
langchain-openai = "^0.2.14"
langchain-anthropic = "^0.3.1"
langgraph = "^0.2.60"
grandalf = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
